Newton

The goal of the small Newton.jl package is to provide a fast and efficient newton-raphson solver for nonlinear equation systems, suitable to be used inside a preformance critical loop. A key feature is that the jacobian at the solution is returned. It is mostly tested for small equations systems (<100 variables). When more fine-grained controlled over algorithms or more iteration information is desired, using NLsolve is recommended.

Usage

Mutating standard array

using Newton

Define a mutating residual function rf!

function rf!(r::Vector, x::Vector)
    return map!(v->(exp(v)-v^2), r, x)
end

Define the unknown array x and preallocate cache cache

x=zeros(5)
cache = NewtonCache(x,rf!)

At the place where we want to solve the problem r=0

x0 = getx(cache)
# Modify x0 as you wish to provide initial guess
x, drdx, converged = newtonsolve!(x0, rf!, cache)

It is not necessary to get x0 from the cache, but this avoids any allocations. This implies that x0 will be aliased to the output, i.e. x0===x after solving.

Using StaticArrays

When using static arrays, the residual function should be non-mutating, i.e.

function rf(x::SVector)
    return exp.(x) - x.^2
end

No cache setup is required for static arrays. Hence, define an initial guess x0 and call the newtonsolve

x0 = zero(SVector{dim})
x, drdx, converged = newtonsolve(x0, rf);

which as in the mutatable array case returns a the solution vector, the jacobian at the solution and a boolean whether the solver converged or not.

Speed comparison

See benchmarks/benchmark.jl, on my laptop the results are

pkg> activate benchmarks/
julia> include("benchmarks/benchmarks.jl");
Benchmark with dim=5
rf (static):           33.099 ns (0 allocations: 0 bytes)
rf (dynamic):          32.931 ns (0 allocations: 0 bytes)
newtonsolve static:    1.000 μs (0 allocations: 0 bytes)
newtonsolve dynamic:   2.400 μs (11 allocations: 1.50 KiB)
nlsolve dynamic:       6.900 μs (58 allocations: 6.23 KiB)

Benchmark with dim=10
rf (static):           61.491 ns (0 allocations: 0 bytes)
rf (dynamic):          66.187 ns (0 allocations: 0 bytes)
newtonsolve static:    4.200 μs (0 allocations: 0 bytes)
newtonsolve dynamic:   5.100 μs (7 allocations: 5.28 KiB)
nlsolve dynamic:       11.400 μs (58 allocations: 12.25 KiB)

Benchmark with dim=20
rf (static):           119.333 ns (0 allocations: 0 bytes)
rf (dynamic):          125.471 ns (0 allocations: 0 bytes)
newtonsolve static:    7.900 μs (16 allocations: 14.81 KiB)
newtonsolve dynamic:   14.600 μs (5 allocations: 4.38 KiB)
nlsolve dynamic:       29.100 μs (62 allocations: 23.39 KiB)

Benchmark with dim=40
rf (static):           265.634 ns (0 allocations: 0 bytes)
rf (dynamic):          251.370 ns (0 allocations: 0 bytes)
newtonsolve static:    38.600 μs (16 allocations: 53.69 KiB)
newtonsolve dynamic:   53.200 μs (5 allocations: 4.38 KiB)
nlsolve dynamic:       83.400 μs (67 allocations: 55.67 KiB)

showing that static arrays are faster than dynamic arrays with newtonsolve and that newtonsolve outperforms nlsolve in these specific cases. (nlsolve does not support StaticArrays.)

Exported API

Newton.NewtonCache — Method

function NewtonCache(x::AbstractVector, rf!)

Create the cache used by the newtonsolve and linsolve!. Only a copy of x will be used.

Newton.getx — Method

getx(cache::NewtonCache)

Extract out the unknown values. This can be used to avoid allocations when solving defining the initial guess.

Newton.newtonsolve — Function

newtonsolve!(x::AbstractVector, drdx::AbstractMatrix, rf!, cache::ResidualCache; tol=1.e-6, maxiter=100)

Solve the nonlinear equation system r(x)=0 using the newton-raphson method. Returns x, drdx, true if converged and x, drdx, false otherwise.

args

x0: Initial guess. (If aliased to getx(cache) it will be mutated.)

rf!: Residual function. Signature rf!(r, x) and mutating the residual r

cache: Optional cache that can be preallocated by calling ResidualCache(x0, rf!)

kwargs

tol=1.e-6: Tolerance on norm(r)

maxiter=100: Maximum number of iterations before no convergence

Newton.newtonsolve — Method

newtonsolve(x0::SVector, rf; tol=1.e-6, maxiter=100)

Solve the nonlinear equation system r(x)=0 using the newton-raphson method. Returns type: (converged, x, drdx), SVector, SMatrix)` where

converged::Bool is true if converged and false otherwise

x::SVector is the solution vector such that r(x)=0

drdx::SMatrix is the jacobian at x

args

x0: Vector of with initial guess for unknowns.

rf: Residual function. Signature r=rf(x::SVector{dim})::SVector{dim}

kwargs

tol=1.e-6: Tolerance on norm(r)

maxiter=100: Maximum number of iterations before no convergence

Internal API

Newton.linsolve! — Method

linsolve!(K::AbstractMatrix, b::AbstractVector, cache::NewtonCache)

Solves the linear equation system Kx=b, mutating both K and b. b is mutated to the solution x