API

Newton.newtonsolve — Function

newtonsolve(x0::AbstractVector, rf!, [cache::NewtonCache]; tol=1.e-6, maxiter=100)

Solve the nonlinear equation (system) r(x)=0 using the newton-raphson method by calling the mutating residual function rf!(r, x), with signature rf!(r::T, x::T)::T where T<:AbstractVector x0 is the initial guess and the optional cache can be preallocated by calling NewtonCache(x0,rf!). Note that x0 is not modified, unless aliased to getx(cache). tol is the tolerance for norm(r) and maxiter the maximum number of iterations.

returns x, drdx, converged::Bool

drdx is the derivative of r wrt. x at the returned x.

newtonsolve(x0::Union{SVector,Number}, rf; tol=1.e-6, maxiter=100)

Solve the nonlinear equation (system) r(x)=0 using the newton-raphson method by calling the residual function r=rf(x), with signature rf(x::T)::T where T<:Union{SVector,Number}. x0 is the initial guess, tol the tolerance form norm(r), and maxiter the maximum number of iterations.

returns: x, drdx, converged::Bool

drdx is the derivative of r wrt. x at the returned x.

Newton.NewtonCache — Type

function NewtonCache(x::AbstractVector, rf!)

Create the cache used by the newtonsolve and linsolve!. Only a copy of x will be used.

Newton.getx — Function

getx(cache::NewtonCache)

Extract out the unknown values. This can be used to avoid allocations when solving defining the initial guess.

Newton.inv! — Function

Newton.inv!(A::Matrix, cache::NewtonCache)

Utilize the LU decomposition from RecursiveFactorization.jl along with the non-exported LinearAlgebra.inv!(::LU) to calculate the inverse of A more efficient than inv(A). Note that A will be used as workspace and values should not be used after calling Newton.inv!

Newton.logging_mode — Function

Newton.logging_mode(; enable=true)

Helper to turn on (enable=true) or off (enable=false) logging of iterations in Newton.jl. Internally, changes the how Newton.@if_logging expr is evaluated: when logging mode is enabled, expr is evaluated, otherwise expr is ignored.

Internal API

Newton.linsolve! — Function

linsolve!(K::AbstractMatrix, b::AbstractVector, cache::NewtonCache)

Solves the linear equation system Kx=b, mutating both K and b. b is mutated to the solution x