Newton

The goal of the small Newton.jl package is to provide a fast and efficient newton-raphson solver for nonlinear equation systems, suitable to be used inside a preformance critical loop. It is mostly tested for small equations systems (<100 variables). When more fine-grained controlled over algorithms or more iteration information is desired, using NLsolve is recommended.

Basic usage

Mutating standard array

using Newton

Define a mutating residual function rf!

nsize=4
(a,b) = [rand(nsize) for _ in 1:2]
function rf!(r, x)
    r .= - a + b.*x + exp.(x)
end

Define initial guess x

x=zeros(nsize)

Preallocate cache

cache = NewtonCache(x,rf!)

Solve the problem r=0

drdx = get_drdx(cache)  # Alternatively drdx=zeros(nsize,nsize), but this allocates 
converged = newtonsolve!(x, drdx, rf!, cache)

Speed and allocation comparison

See benchmarks/benchmark.jl, on my laptop the results are

include("benchmarks/benchmark.jl")
@btime rf!
  200.000 ns (0 allocations: 0 bytes)
@btime newtonsolve!
  29.400 μs (7 allocations: 6.12 KiB)
@btime nlsolve
  51.500 μs (74 allocations: 40.98 KiB)
Benchmark (dim=20) complete

showing that newtonsolve! is approximately 1.75 times faster than the basic usage of nlsolve for this particular case.

Using StaticArrays

using Newton

Define a non-mutating residual function

function rf(x::SVector)
    return exp.(x) - x.^2
end

Provide an initial guess

x_s = zero(SVector{dim})

Solve the non-linear equation system

converged, drdx = newtonsolve($x_s, $rf);

Speed comparison

See benchmarks/benchmark_static.jl, on my laptop the results are

include("benchmarks/benchmark_static.jl")
Benchmark with dim=5
rf (static):           27.163 ns (0 allocations: 0 bytes)
rf (dynamic):          30.050 ns (0 allocations: 0 bytes)
newtonsolve static:    958.333 ns (0 allocations: 0 bytes)
newtonsolve dynamic:   2.600 μs (9 allocations: 1.22 KiB)
nlsolve dynamic:       6.600 μs (58 allocations: 6.23 KiB)

Benchmark with dim=10
rf (static):           56.504 ns (0 allocations: 0 bytes)
rf (dynamic):          55.793 ns (0 allocations: 0 bytes)
newtonsolve static:    4.037 μs (0 allocations: 0 bytes)
newtonsolve dynamic:   5.400 μs (5 allocations: 4.38 KiB)
nlsolve dynamic:       10.400 μs (58 allocations: 12.25 KiB)

Benchmark with dim=20
rf (static):           105.443 ns (0 allocations: 0 bytes)
rf (dynamic):          109.257 ns (0 allocations: 0 bytes)
newtonsolve static:    8.433 μs (16 allocations: 14.81 KiB)
newtonsolve dynamic:   14.800 μs (5 allocations: 4.38 KiB)
nlsolve dynamic:       25.800 μs (62 allocations: 23.39 KiB)

showing that using StaticArrays will be significantly faster with newtonsolve. (nlsolve does not support StaticArrays.)

Exported API

Newton.NewtonCache — Method

function NewtonCache(x::AbstractVector, rf!)

Create the cache used by the newtonsolve! and linsolve! to find x such that rf!(r,x) yields r=0.

Newton.get_drdx — Method

get_drdx(cache::NewtonCache) = DiffResults.jacobian(cache.result)

Newton.newtonsolve! — Function

newtonsolve!(x::AbstractVector, drdx::AbstractMatrix, rf!, cache::ResidualCache; tol=1.e-6, max_iter=100)

Solve the nonlinear equation system r(x)=0 using the newton-raphson method. Returns true if converged and false otherwise.

args

x: Vector of unknowns. Provide as initial guess, mutated to solution.

drdx: Jacobian matrix. Only provided as preallocation. Can be aliased to DiffResults.jacobian(cache.result)

rf!: Residual function. Signature rf!(r, x) and mutating the residual r

cache: Optional cache that can be preallocated by calling ResidualCache(x, rf!)

kwargs

tol=1.e-6: Tolerance on norm(r)

maxiter=100: Maximum number of iterations before no convergence

Newton.newtonsolve — Method

newtonsolve(x::SVector, rf; tol=1.e-6, max_iter=100)

Solve the nonlinear equation system r(x)=0 using the newton-raphson method. Return type: (Bool, SMatrix) where the Bool is trueif converged andfalseotherwise TheSMatrixis the jacobiandrdxwherer(x)=0`

args

x: Vector of unknowns. Provide as initial guess, mutated to solution.

rf: Residual function. Signature r=rf(x::SVector{dim})::SVector{dim}

kwargs

tol=1.e-6: Tolerance on norm(r)

maxiter=100: Maximum number of iterations before no convergence

Internal API

Newton.linsolve! — Method

linsolve!(K::AbstractMatrix, b::AbstractVector, cache::NewtonCache)

Solves the linear equation system Kx=b, mutating both K and b. b is mutated to the solution x

Newton.linsolve — Method

linsolve(drdx::SMatrix{dim,dim}, r::SVector{dim}) where{dim}

Solves the linear equation system drdx*x=r without mutating and returns the solution x